Make Ascribe behaviour user-modifiable and easily extendable
This change adds a 'b:attributes' variable to each buffer. This variable
contains the list of attributes reported by Git for that file.

Also introduced is the 'g:ascribe_handlers' dictionary. This dictionary
defines the attributes which should be checked (as keys) and maps them
to an appropriate attribute handler (given as a 'Funcref').